requirements.txt
absl-py==0.12.0
pandas==1.2.4
Pillow==8.2.0
torch==1.8.0
torchnet==0.0.4
torchvision==0.9.0
